Add refactor: moved class into its own file
