+++ changed.txt
@@ -1,1 +1,1 @@
-THIS IS A TEST.
+THIS IS A MESSAGE.
